OCD: Update screen recording vectors & simplify strings
Previously the tile strings would indicate which quality setting the user
had the tile set to. This was fine except that the strings were long, and
depending on the current QS layout/screen desity, the tite titles had to
auto scroll in order to be read. Being the only tile with this happening,
it just looked kind of janky. This uses the vectors to tell the user whether
they're on 480p, 720p, or 720p+, while the tile title just says "Screenrecord".

Change-Id: I1f7c5071192cf33fd9527323a3769db23d395d54
Signed-off-by: calebcabob <calphonic@gmail.com>